heartform/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
checkbox-support==0.22
Click==7.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
dell-recovery==0.0.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.7
django-crispy-forms==1.8.0
feedparser==5.2.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
guacamole==0.9.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==2.1.1
netifaces==0.10.4
numexpr==2.6.4
numpy==1.17.3
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
onboard==1.4.1
padme==1.1.1
pandas==0.22.0
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
plainbox==0.25
progressbar==2.3
protobuf==3.0.0
psycopg2==2.7.4
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.21.3
scipy==1.3.1
scour==0.36
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tables==3.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.16.0
xdiagnose==3.8.8
xkit==0.0.0
XlsxWriter==0.9.6
zope.interface==4.3.2
